+++ StrategyTemplate/ForYou/Indicators/Groups/ICycles.cs
@@ -16,7 +16,7 @@ namespace StrategyTemplate.ForYou.Indicators.Groups
         /// <param name="unitType"></param>
         /// <param name="historiesLength"></param>
         /// <returns></returns>
-        ITaLibOutput<IList<double>> HilbertTransformDominantCyclePeriod(Frequency unitType, int historiesLength);
+        ITaLibOutput<IList<decimal>> HilbertTransformDominantCyclePeriod(Frequency unitType, int historiesLength);
 
         /// <summary>
         /// NOTE: The HT_DCPHASE function has an unstable period.
@@ -25,7 +25,7 @@ namespace StrategyTemplate.ForYou.Indicators.Groups
         /// <param name="unitType"></param>
         /// <param name="historiesLength"></param>
         /// <returns></returns>
-        ITaLibOutput<IList<double>> HilbertTransformDominantCyclePhase(Frequency unitType, int historiesLength);
+        ITaLibOutput<IList<decimal>> HilbertTransformDominantCyclePhase(Frequency unitType, int historiesLength);
 
         /// <summary>
         /// NOTE: The HT_PHASOR function has an unstable period.
@@ -34,7 +34,7 @@ namespace StrategyTemplate.ForYou.Indicators.Groups
         /// <param name="unitType"></param>
         /// <param name="historiesLength"></param>
         /// <returns></returns>
-        ITaLibOutput<(IList<double> InPhase, IList<double> Quadrature)> HilbertTransformPhasorComponents(Frequency unitType, int historiesLength);
+        ITaLibOutput<(IList<decimal> InPhase, IList<decimal> Quadrature)> HilbertTransformPhasorComponents(Frequency unitType, int historiesLength);
 
         /// <summary>
         /// NOTE: The HT_SINE function has an unstable period.
@@ -43,7 +43,7 @@ namespace StrategyTemplate.ForYou.Indicators.Groups
         /// <param name="unitType"></param>
         /// <param name="historiesLength"></param>
         /// <returns></returns>
-        ITaLibOutput<(IList<double> Sine, IList<double> LeadSine)> HilbertTransformSineWave(Frequency unitType, int historiesLength);
+        ITaLibOutput<(IList<decimal> Sine, IList<decimal> LeadSine)> HilbertTransformSineWave(Frequency unitType, int historiesLength);
 
         /// <summary>
         /// NOTE: The HT_TRENDMODE function has an unstable period.
